use tomato url from omdbapi instead of guessing it
thanks to a google search leading to
https://greasyfork.org/en/scripts/7584-rotten-tomatoes-link-on-imdb/code
we found out that there's a parameter `tomatoes=true` to omdbapi which
results in the tomato url to be included in the api's response. this
parameter is not documented on omdbapi.
we now don't need to try and guess the tomato url anymore,