cli: support json output
I intend to replace the svelte web app by a rails app, that will call
out to the cli in order to get the data. The whole SPA + FAAS setup was
a nice experiment, but in the long run it's a pain in the ass to
maintain. Production is also a lot more involved, fails more (especially
FAAS) and costs more.