search by normalized and original title
The searchable title is normalized to be able to search for the term
independent of accents etc, e.g. "pokemon" for "Pokémon". There are also
records where there's not "en" title. In that case we fall back to
"en-us". Additionally we now also search for the original title in the
original language.