move from omdb to full wikidata
This drops the dependency from OMDb API and uses wikidata for
everything. The crystal (CLI) parts are ported to ruby to accomodate the
web application. At some later point I'll either deprecate the CLI or
rewrite it to use the web application through an API instead.

With this we can also use the `wiki_id` as the primary internal id, and
don't have to rely on the `imdb_id` anymore. This also means that we can
search for all movies, not only for the ones with an IMDb reference.
The autocomplete now also redirects directly to the target show path
instead of going through the search.