remove the "Look up" search button
The search suggestions are the primary means of searching and
navigating to a movie. Using the input field directly is somewhat of a
side show, and should not be allowed.
At a later point I'll add "See all results" to the search suggestion to
be able to show a full search results page.