prevent infinite loop on fetching/storing failure
When a movie's score is empty we wouldn't store the scores in redis.
That would lead to an infinite loop in the `MovieController`.

Contrary to the previous commit's claim it still makes sense to have a
hint in the UI when fetching of a score fails.

We now show a disclaimer that not all score could be fetch, and persist
the score in redis in any case. In the event of fetching failure we set
a small TTL to allow re-fetching the scores in the next try, while still
preventing an infinite loop.