deterministic records ordering
The order by rank and release_date was not a unique sort key. The result
was non detarministcly ordered. This lead to the result items swapping
on "Show more results".

We now also use sql union instead of combining 2 separate queries in
memory. This makes it easier to reuse the order sql.

Since the union of the two result sets would not eliminate duplicates
since the rank would render the records non distinct we need to select
double the records and then make the records unique in memory.