reject too short or unspecific search queries
When entering e.g. "the" the query would take several seconds to
complete. From a user perspective the search result is also undefined,
which is why it makes more sense to reject such terms and show a message
to the user instead.

We now show "2 more characters…", "1 more character…" in case the search
term isn't long enough, and "Not specific enough…" when the search is
either "the" or times out. For now I set the query timeout for search to
200ms. Every query taking longer than that is rejected.