add share button on mobile
Ever since I changed the PWA display mode to standalone it's not
possible anymore to share the url.
A better option is to use the Web Share API anyway. This will trigger
the mobile native share dialogue.

This also fixes the social media meta tags to contain absolute image
urls.